Port tests to e2e node
