+++ test/e2e/pods.go
@@ -20,26 +20,19 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"strconv"
 	"strings"
 	"time"
 
 	"golang.org/x/net/websocket"
 
 	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/api/errors"
-	"k8s.io/kubernetes/pkg/api/resource"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
 	"k8s.io/kubernetes/pkg/kubelet"
-	"k8s.io/kubernetes/pkg/labels"
 	"k8s.io/kubernetes/pkg/util"
 	"k8s.io/kubernetes/pkg/util/intstr"
-	"k8s.io/kubernetes/pkg/util/wait"
-	"k8s.io/kubernetes/pkg/watch"
 	"k8s.io/kubernetes/test/e2e/framework"
 
 	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 )
 
 const (
@@ -110,665 +103,8 @@ func runLivenessTest(c *client.Client, ns string, podDescr *api.Pod, expectNumRe
 	}
 }
 
-// testHostIP tests that a pod gets a host IP
-func testHostIP(c *client.Client, ns string, pod *api.Pod) {
-	podClient := c.Pods(ns)
-	By("creating pod")
-	defer podClient.Delete(pod.Name, api.NewDeleteOptions(0))
-	if _, err := podClient.Create(pod); err != nil {
-		framework.Failf("Failed to create pod: %v", err)
-	}
-	By("ensuring that pod is running and has a hostIP")
-	// Wait for the pods to enter the running state. Waiting loops until the pods
-	// are running so non-running pods cause a timeout for this test.
-	err := framework.WaitForPodRunningInNamespace(c, pod.Name, ns)
-	Expect(err).NotTo(HaveOccurred())
-	// Try to make sure we get a hostIP for each pod.
-	hostIPTimeout := 2 * time.Minute
-	t := time.Now()
-	for {
-		p, err := podClient.Get(pod.Name)
-		Expect(err).NotTo(HaveOccurred())
-		if p.Status.HostIP != "" {
-			framework.Logf("Pod %s has hostIP: %s", p.Name, p.Status.HostIP)
-			break
-		}
-		if time.Since(t) >= hostIPTimeout {
-			framework.Failf("Gave up waiting for hostIP of pod %s after %v seconds",
-				p.Name, time.Since(t).Seconds())
-		}
-		framework.Logf("Retrying to get the hostIP of pod %s", p.Name)
-		time.Sleep(5 * time.Second)
-	}
-}
-
-func runPodFromStruct(f *framework.Framework, pod *api.Pod) {
-	By("submitting the pod to kubernetes")
-
-	podClient := f.Client.Pods(f.Namespace.Name)
-	pod, err := podClient.Create(pod)
-	if err != nil {
-		framework.Failf("Failed to create pod: %v", err)
-	}
-
-	framework.ExpectNoError(f.WaitForPodRunning(pod.Name))
-
-	By("verifying the pod is in kubernetes")
-	pod, err = podClient.Get(pod.Name)
-	if err != nil {
-		framework.Failf("failed to get pod: %v", err)
-	}
-}
-
-func startPodAndGetBackOffs(f *framework.Framework, pod *api.Pod, podName string, containerName string, sleepAmount time.Duration) (time.Duration, time.Duration) {
-	runPodFromStruct(f, pod)
-	time.Sleep(sleepAmount)
-
-	By("getting restart delay-0")
-	_, err := getRestartDelay(f.Client, pod, f.Namespace.Name, podName, containerName)
-	if err != nil {
-		framework.Failf("timed out waiting for container restart in pod=%s/%s", podName, containerName)
-	}
-
-	By("getting restart delay-1")
-	delay1, err := getRestartDelay(f.Client, pod, f.Namespace.Name, podName, containerName)
-	if err != nil {
-		framework.Failf("timed out waiting for container restart in pod=%s/%s", podName, containerName)
-	}
-
-	By("getting restart delay-2")
-	delay2, err := getRestartDelay(f.Client, pod, f.Namespace.Name, podName, containerName)
-	if err != nil {
-		framework.Failf("timed out waiting for container restart in pod=%s/%s", podName, containerName)
-	}
-	return delay1, delay2
-}
-
-func getRestartDelay(c *client.Client, pod *api.Pod, ns string, name string, containerName string) (time.Duration, error) {
-	beginTime := time.Now()
-	for time.Since(beginTime) < (2 * maxBackOffTolerance) { // may just miss the 1st MaxContainerBackOff delay
-		time.Sleep(time.Second)
-		pod, err := c.Pods(ns).Get(name)
-		framework.ExpectNoError(err, fmt.Sprintf("getting pod %s", name))
-		status, ok := api.GetContainerStatus(pod.Status.ContainerStatuses, containerName)
-		if !ok {
-			framework.Logf("getRestartDelay: status missing")
-			continue
-		}
-
-		if status.State.Waiting == nil && status.State.Running != nil && status.LastTerminationState.Terminated != nil && status.State.Running.StartedAt.Time.After(beginTime) {
-			startedAt := status.State.Running.StartedAt.Time
-			finishedAt := status.LastTerminationState.Terminated.FinishedAt.Time
-			framework.Logf("getRestartDelay: restartCount = %d, finishedAt=%s restartedAt=%s (%s)", status.RestartCount, finishedAt, startedAt, startedAt.Sub(finishedAt))
-			return startedAt.Sub(finishedAt), nil
-		}
-	}
-	return 0, fmt.Errorf("timeout getting pod restart delay")
-}
-
 var _ = framework.KubeDescribe("Pods", func() {
 	f := framework.NewDefaultFramework("pods")
-
-	It("should get a host IP [Conformance]", func() {
-		name := "pod-hostip-" + string(util.NewUUID())
-		testHostIP(f.Client, f.Namespace.Name, &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name: name,
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:  "test",
-						Image: "gcr.io/google_containers/pause-amd64:3.0",
-					},
-				},
-			},
-		})
-	})
-
-	It("should be schedule with cpu and memory limits [Conformance]", func() {
-		podClient := f.Client.Pods(f.Namespace.Name)
-
-		By("creating the pod")
-		name := "pod-update-" + string(util.NewUUID())
-		value := strconv.Itoa(time.Now().Nanosecond())
-		pod := &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name: name,
-				Labels: map[string]string{
-					"name": "foo",
-					"time": value,
-				},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:  "nginx",
-						Image: "gcr.io/google_containers/pause-amd64:3.0",
-						Resources: api.ResourceRequirements{
-							Limits: api.ResourceList{
-								api.ResourceCPU:    *resource.NewMilliQuantity(100, resource.DecimalSI),
-								api.ResourceMemory: *resource.NewQuantity(10*1024*1024, resource.DecimalSI),
-							},
-						},
-					},
-				},
-			},
-		}
-		defer podClient.Delete(pod.Name, nil)
-		_, err := podClient.Create(pod)
-		if err != nil {
-			framework.Failf("Error creating a pod: %v", err)
-		}
-		framework.ExpectNoError(f.WaitForPodRunning(pod.Name))
-	})
-
-	It("should be submitted and removed [Conformance]", func() {
-		podClient := f.Client.Pods(f.Namespace.Name)
-
-		By("creating the pod")
-		name := "pod-update-" + string(util.NewUUID())
-		value := strconv.Itoa(time.Now().Nanosecond())
-		pod := &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name: name,
-				Labels: map[string]string{
-					"name": "foo",
-					"time": value,
-				},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:  "nginx",
-						Image: "gcr.io/google_containers/nginx:1.7.9",
-						Ports: []api.ContainerPort{{ContainerPort: 80}},
-						LivenessProbe: &api.Probe{
-							Handler: api.Handler{
-								HTTPGet: &api.HTTPGetAction{
-									Path: "/index.html",
-									Port: intstr.FromInt(8080),
-								},
-							},
-							InitialDelaySeconds: 30,
-						},
-					},
-				},
-			},
-		}
-
-		By("setting up watch")
-		selector := labels.SelectorFromSet(labels.Set(map[string]string{"time": value}))
-		options := api.ListOptions{LabelSelector: selector}
-		pods, err := podClient.List(options)
-		if err != nil {
-			framework.Failf("Failed to query for pods: %v", err)
-		}
-		Expect(len(pods.Items)).To(Equal(0))
-		options = api.ListOptions{
-			LabelSelector:   selector,
-			ResourceVersion: pods.ListMeta.ResourceVersion,
-		}
-		w, err := podClient.Watch(options)
-		if err != nil {
-			framework.Failf("Failed to set up watch: %v", err)
-		}
-
-		By("submitting the pod to kubernetes")
-		// We call defer here in case there is a problem with
-		// the test so we can ensure that we clean up after
-		// ourselves
-		defer podClient.Delete(pod.Name, api.NewDeleteOptions(0))
-		_, err = podClient.Create(pod)
-		if err != nil {
-			framework.Failf("Failed to create pod: %v", err)
-		}
-
-		By("verifying the pod is in kubernetes")
-		selector = labels.SelectorFromSet(labels.Set(map[string]string{"time": value}))
-		options = api.ListOptions{LabelSelector: selector}
-		pods, err = podClient.List(options)
-		if err != nil {
-			framework.Failf("Failed to query for pods: %v", err)
-		}
-		Expect(len(pods.Items)).To(Equal(1))
-
-		By("verifying pod creation was observed")
-		select {
-		case event, _ := <-w.ResultChan():
-			if event.Type != watch.Added {
-				framework.Failf("Failed to observe pod creation: %v", event)
-			}
-		case <-time.After(framework.PodStartTimeout):
-			Fail("Timeout while waiting for pod creation")
-		}
-
-		// We need to wait for the pod to be scheduled, otherwise the deletion
-		// will be carried out immediately rather than gracefully.
-		framework.ExpectNoError(f.WaitForPodRunning(pod.Name))
-
-		By("deleting the pod gracefully")
-		if err := podClient.Delete(pod.Name, api.NewDeleteOptions(30)); err != nil {
-			framework.Failf("Failed to delete pod: %v", err)
-		}
-
-		By("verifying the kubelet observed the termination notice")
-		pod, err = podClient.Get(pod.Name)
-		Expect(wait.Poll(time.Second*5, time.Second*30, func() (bool, error) {
-			podList, err := framework.GetKubeletPods(f.Client, pod.Spec.NodeName)
-			if err != nil {
-				framework.Logf("Unable to retrieve kubelet pods for node %v: %v", pod.Spec.NodeName, err)
-				return false, nil
-			}
-			for _, kubeletPod := range podList.Items {
-				if pod.Name != kubeletPod.Name {
-					continue
-				}
-				if kubeletPod.ObjectMeta.DeletionTimestamp == nil {
-					framework.Logf("deletion has not yet been observed")
-					return false, nil
-				}
-				return true, nil
-			}
-			framework.Logf("no pod exists with the name we were looking for, assuming the termination request was observed and completed")
-			return true, nil
-		})).NotTo(HaveOccurred(), "kubelet never observed the termination notice")
-
-		By("verifying pod deletion was observed")
-		deleted := false
-		timeout := false
-		var lastPod *api.Pod
-		timer := time.After(30 * time.Second)
-		for !deleted && !timeout {
-			select {
-			case event, _ := <-w.ResultChan():
-				if event.Type == watch.Deleted {
-					lastPod = event.Object.(*api.Pod)
-					deleted = true
-				}
-			case <-timer:
-				timeout = true
-			}
-		}
-		if !deleted {
-			Fail("Failed to observe pod deletion")
-		}
-
-		Expect(lastPod.DeletionTimestamp).ToNot(BeNil())
-		Expect(lastPod.Spec.TerminationGracePeriodSeconds).ToNot(BeZero())
-
-		selector = labels.SelectorFromSet(labels.Set(map[string]string{"time": value}))
-		options = api.ListOptions{LabelSelector: selector}
-		pods, err = podClient.List(options)
-		if err != nil {
-			Fail(fmt.Sprintf("Failed to list pods to verify deletion: %v", err))
-		}
-		Expect(len(pods.Items)).To(Equal(0))
-	})
-
-	It("should be updated [Conformance]", func() {
-		podClient := f.Client.Pods(f.Namespace.Name)
-
-		By("creating the pod")
-		name := "pod-update-" + string(util.NewUUID())
-		value := strconv.Itoa(time.Now().Nanosecond())
-		pod := &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name: name,
-				Labels: map[string]string{
-					"name": "foo",
-					"time": value,
-				},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:  "nginx",
-						Image: "gcr.io/google_containers/nginx:1.7.9",
-						Ports: []api.ContainerPort{{ContainerPort: 80}},
-						LivenessProbe: &api.Probe{
-							Handler: api.Handler{
-								HTTPGet: &api.HTTPGetAction{
-									Path: "/index.html",
-									Port: intstr.FromInt(8080),
-								},
-							},
-							InitialDelaySeconds: 30,
-						},
-					},
-				},
-			},
-		}
-
-		By("submitting the pod to kubernetes")
-		defer func() {
-			By("deleting the pod")
-			podClient.Delete(pod.Name, api.NewDeleteOptions(0))
-		}()
-		pod, err := podClient.Create(pod)
-		if err != nil {
-			framework.Failf("Failed to create pod: %v", err)
-		}
-
-		framework.ExpectNoError(f.WaitForPodRunning(pod.Name))
-
-		By("verifying the pod is in kubernetes")
-		selector := labels.SelectorFromSet(labels.Set(map[string]string{"time": value}))
-		options := api.ListOptions{LabelSelector: selector}
-		pods, err := podClient.List(options)
-		Expect(len(pods.Items)).To(Equal(1))
-
-		// Standard get, update retry loop
-		framework.ExpectNoError(wait.Poll(time.Millisecond*500, time.Second*30, func() (bool, error) {
-			By("updating the pod")
-			value = strconv.Itoa(time.Now().Nanosecond())
-			if pod == nil { // on retries we need to re-get
-				pod, err = podClient.Get(name)
-				if err != nil {
-					return false, fmt.Errorf("failed to get pod: %v", err)
-				}
-			}
-			pod.Labels["time"] = value
-			pod, err = podClient.Update(pod)
-			if err == nil {
-				framework.Logf("Successfully updated pod")
-				return true, nil
-			}
-			if errors.IsConflict(err) {
-				framework.Logf("Conflicting update to pod, re-get and re-update: %v", err)
-				pod = nil // re-get it when we retry
-				return false, nil
-			}
-			return false, fmt.Errorf("failed to update pod: %v", err)
-		}))
-
-		framework.ExpectNoError(f.WaitForPodRunning(pod.Name))
-
-		By("verifying the updated pod is in kubernetes")
-		selector = labels.SelectorFromSet(labels.Set(map[string]string{"time": value}))
-		options = api.ListOptions{LabelSelector: selector}
-		pods, err = podClient.List(options)
-		Expect(len(pods.Items)).To(Equal(1))
-		framework.Logf("Pod update OK")
-	})
-
-	It("should allow activeDeadlineSeconds to be updated [Conformance]", func() {
-		podClient := f.Client.Pods(f.Namespace.Name)
-
-		By("creating the pod")
-		name := "pod-update-activedeadlineseconds-" + string(util.NewUUID())
-		value := strconv.Itoa(time.Now().Nanosecond())
-		pod := &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name: name,
-				Labels: map[string]string{
-					"name": "foo",
-					"time": value,
-				},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:  "nginx",
-						Image: "gcr.io/google_containers/nginx:1.7.9",
-						Ports: []api.ContainerPort{{ContainerPort: 80}},
-						LivenessProbe: &api.Probe{
-							Handler: api.Handler{
-								HTTPGet: &api.HTTPGetAction{
-									Path: "/index.html",
-									Port: intstr.FromInt(8080),
-								},
-							},
-							InitialDelaySeconds: 30,
-						},
-					},
-				},
-			},
-		}
-
-		By("submitting the pod to kubernetes")
-		defer func() {
-			By("deleting the pod")
-			podClient.Delete(pod.Name, api.NewDeleteOptions(0))
-		}()
-		pod, err := podClient.Create(pod)
-		if err != nil {
-			framework.Failf("Failed to create pod: %v", err)
-		}
-
-		framework.ExpectNoError(f.WaitForPodRunning(pod.Name))
-
-		By("verifying the pod is in kubernetes")
-		selector := labels.SelectorFromSet(labels.Set(map[string]string{"time": value}))
-		options := api.ListOptions{LabelSelector: selector}
-		pods, err := podClient.List(options)
-		Expect(len(pods.Items)).To(Equal(1))
-
-		// Standard get, update retry loop
-		framework.ExpectNoError(wait.Poll(time.Millisecond*500, time.Second*30, func() (bool, error) {
-			By("updating the pod")
-			value = strconv.Itoa(time.Now().Nanosecond())
-			if pod == nil { // on retries we need to re-get
-				pod, err = podClient.Get(name)
-				if err != nil {
-					return false, fmt.Errorf("failed to get pod: %v", err)
-				}
-			}
-			newDeadline := int64(5)
-			pod.Spec.ActiveDeadlineSeconds = &newDeadline
-			pod, err = podClient.Update(pod)
-			if err == nil {
-				framework.Logf("Successfully updated pod")
-				return true, nil
-			}
-			if errors.IsConflict(err) {
-				framework.Logf("Conflicting update to pod, re-get and re-update: %v", err)
-				pod = nil // re-get it when we retry
-				return false, nil
-			}
-			return false, fmt.Errorf("failed to update pod: %v", err)
-		}))
-
-		framework.ExpectNoError(f.WaitForPodTerminated(pod.Name, "DeadlineExceeded"))
-	})
-
-	It("should contain environment variables for services [Conformance]", func() {
-		// Make a pod that will be a service.
-		// This pod serves its hostname via HTTP.
-		serverName := "server-envvars-" + string(util.NewUUID())
-		serverPod := &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name:   serverName,
-				Labels: map[string]string{"name": serverName},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:  "srv",
-						Image: "gcr.io/google_containers/serve_hostname:v1.4",
-						Ports: []api.ContainerPort{{ContainerPort: 9376}},
-					},
-				},
-			},
-		}
-		defer f.Client.Pods(f.Namespace.Name).Delete(serverPod.Name, api.NewDeleteOptions(0))
-		_, err := f.Client.Pods(f.Namespace.Name).Create(serverPod)
-		if err != nil {
-			framework.Failf("Failed to create serverPod: %v", err)
-		}
-		framework.ExpectNoError(f.WaitForPodRunning(serverPod.Name))
-
-		// This service exposes port 8080 of the test pod as a service on port 8765
-		// TODO(filbranden): We would like to use a unique service name such as:
-		//   svcName := "svc-envvars-" + randomSuffix()
-		// However, that affects the name of the environment variables which are the capitalized
-		// service name, so that breaks this test.  One possibility is to tweak the variable names
-		// to match the service.  Another is to rethink environment variable names and possibly
-		// allow overriding the prefix in the service manifest.
-		svcName := "fooservice"
-		svc := &api.Service{
-			ObjectMeta: api.ObjectMeta{
-				Name: svcName,
-				Labels: map[string]string{
-					"name": svcName,
-				},
-			},
-			Spec: api.ServiceSpec{
-				Ports: []api.ServicePort{{
-					Port:       8765,
-					TargetPort: intstr.FromInt(8080),
-				}},
-				Selector: map[string]string{
-					"name": serverName,
-				},
-			},
-		}
-		defer f.Client.Services(f.Namespace.Name).Delete(svc.Name)
-		_, err = f.Client.Services(f.Namespace.Name).Create(svc)
-		if err != nil {
-			framework.Failf("Failed to create service: %v", err)
-		}
-
-		// Make a client pod that verifies that it has the service environment variables.
-		podName := "client-envvars-" + string(util.NewUUID())
-		pod := &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name:   podName,
-				Labels: map[string]string{"name": podName},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:    "env3cont",
-						Image:   "gcr.io/google_containers/busybox:1.24",
-						Command: []string{"sh", "-c", "env"},
-					},
-				},
-				RestartPolicy: api.RestartPolicyNever,
-			},
-		}
-
-		f.TestContainerOutput("service env", pod, 0, []string{
-			"FOOSERVICE_SERVICE_HOST=",
-			"FOOSERVICE_SERVICE_PORT=",
-			"FOOSERVICE_PORT=",
-			"FOOSERVICE_PORT_8765_TCP_PORT=",
-			"FOOSERVICE_PORT_8765_TCP_PROTO=",
-			"FOOSERVICE_PORT_8765_TCP=",
-			"FOOSERVICE_PORT_8765_TCP_ADDR=",
-		})
-	})
-
-	It("should be restarted with a docker exec \"cat /tmp/health\" liveness probe [Conformance]", func() {
-		runLivenessTest(f.Client, f.Namespace.Name, &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name:   "liveness-exec",
-				Labels: map[string]string{"test": "liveness"},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:    "liveness",
-						Image:   "gcr.io/google_containers/busybox:1.24",
-						Command: []string{"/bin/sh", "-c", "echo ok >/tmp/health; sleep 10; rm -rf /tmp/health; sleep 600"},
-						LivenessProbe: &api.Probe{
-							Handler: api.Handler{
-								Exec: &api.ExecAction{
-									Command: []string{"cat", "/tmp/health"},
-								},
-							},
-							InitialDelaySeconds: 15,
-							FailureThreshold:    1,
-						},
-					},
-				},
-			},
-		}, 1, defaultObservationTimeout)
-	})
-
-	It("should *not* be restarted with a docker exec \"cat /tmp/health\" liveness probe [Conformance]", func() {
-		runLivenessTest(f.Client, f.Namespace.Name, &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name:   "liveness-exec",
-				Labels: map[string]string{"test": "liveness"},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:    "liveness",
-						Image:   "gcr.io/google_containers/busybox:1.24",
-						Command: []string{"/bin/sh", "-c", "echo ok >/tmp/health; sleep 600"},
-						LivenessProbe: &api.Probe{
-							Handler: api.Handler{
-								Exec: &api.ExecAction{
-									Command: []string{"cat", "/tmp/health"},
-								},
-							},
-							InitialDelaySeconds: 15,
-							FailureThreshold:    1,
-						},
-					},
-				},
-			},
-		}, 0, defaultObservationTimeout)
-	})
-
-	It("should be restarted with a /healthz http liveness probe [Conformance]", func() {
-		runLivenessTest(f.Client, f.Namespace.Name, &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name:   "liveness-http",
-				Labels: map[string]string{"test": "liveness"},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:    "liveness",
-						Image:   "gcr.io/google_containers/liveness:e2e",
-						Command: []string{"/server"},
-						LivenessProbe: &api.Probe{
-							Handler: api.Handler{
-								HTTPGet: &api.HTTPGetAction{
-									Path: "/healthz",
-									Port: intstr.FromInt(8080),
-								},
-							},
-							InitialDelaySeconds: 15,
-							FailureThreshold:    1,
-						},
-					},
-				},
-			},
-		}, 1, defaultObservationTimeout)
-	})
-
-	// Slow by design (5 min)
-	It("should have monotonically increasing restart count [Conformance] [Slow]", func() {
-		runLivenessTest(f.Client, f.Namespace.Name, &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name:   "liveness-http",
-				Labels: map[string]string{"test": "liveness"},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:    "liveness",
-						Image:   "gcr.io/google_containers/liveness:e2e",
-						Command: []string{"/server"},
-						LivenessProbe: &api.Probe{
-							Handler: api.Handler{
-								HTTPGet: &api.HTTPGetAction{
-									Path: "/healthz",
-									Port: intstr.FromInt(8080),
-								},
-							},
-							InitialDelaySeconds: 5,
-							FailureThreshold:    1,
-						},
-					},
-				},
-			},
-		}, 5, time.Minute*5)
-	})
 
 	It("should *not* be restarted with a /healthz http liveness probe [Conformance]", func() {
 		runLivenessTest(f.Client, f.Namespace.Name, &api.Pod{
@@ -950,118 +286,6 @@ var _ = framework.KubeDescribe("Pods", func() {
 		}
 		if buf.String() != "container is alive\n" {
 			framework.Failf("Unexpected websocket logs:\n%s", buf.String())
-		}
-	})
-
-	It("should have their auto-restart back-off timer reset on image update [Slow]", func() {
-		podName := "pod-back-off-image"
-		containerName := "back-off"
-		podClient := f.Client.Pods(f.Namespace.Name)
-		pod := &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name:   podName,
-				Labels: map[string]string{"test": "back-off-image"},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:    containerName,
-						Image:   "gcr.io/google_containers/busybox:1.24",
-						Command: []string{"/bin/sh", "-c", "sleep 5", "/crash/missing"},
-					},
-				},
-			},
-		}
-
-		defer func() {
-			By("deleting the pod")
-			podClient.Delete(pod.Name, api.NewDeleteOptions(0))
-		}()
-
-		delay1, delay2 := startPodAndGetBackOffs(f, pod, podName, containerName, buildBackOffDuration)
-
-		By("updating the image")
-		pod, err := podClient.Get(pod.Name)
-		if err != nil {
-			framework.Failf("failed to get pod: %v", err)
-		}
-		pod.Spec.Containers[0].Image = "gcr.io/google_containers/nginx:1.7.9"
-		pod, err = podClient.Update(pod)
-		if err != nil {
-			framework.Failf("error updating pod=%s/%s %v", podName, containerName, err)
-		}
-		time.Sleep(syncLoopFrequency)
-		framework.ExpectNoError(f.WaitForPodRunning(pod.Name))
-
-		By("get restart delay after image update")
-		delayAfterUpdate, err := getRestartDelay(f.Client, pod, f.Namespace.Name, podName, containerName)
-		if err != nil {
-			framework.Failf("timed out waiting for container restart in pod=%s/%s", podName, containerName)
-		}
-
-		if delayAfterUpdate > 2*delay2 || delayAfterUpdate > 2*delay1 {
-			framework.Failf("updating image did not reset the back-off value in pod=%s/%s d3=%s d2=%s d1=%s", podName, containerName, delayAfterUpdate, delay1, delay2)
-		}
-	})
-
-	// Slow issue #19027 (20 mins)
-	It("should cap back-off at MaxContainerBackOff [Slow]", func() {
-		podClient := f.Client.Pods(f.Namespace.Name)
-		podName := "back-off-cap"
-		containerName := "back-off-cap"
-		pod := &api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name:   podName,
-				Labels: map[string]string{"test": "liveness"},
-			},
-			Spec: api.PodSpec{
-				Containers: []api.Container{
-					{
-						Name:    containerName,
-						Image:   "gcr.io/google_containers/busybox:1.24",
-						Command: []string{"/bin/sh", "-c", "sleep 5", "/crash/missing"},
-					},
-				},
-			},
-		}
-
-		defer func() {
-			By("deleting the pod")
-			podClient.Delete(pod.Name, api.NewDeleteOptions(0))
-		}()
-
-		runPodFromStruct(f, pod)
-		time.Sleep(2 * kubelet.MaxContainerBackOff) // it takes slightly more than 2*x to get to a back-off of x
-
-		// wait for a delay == capped delay of MaxContainerBackOff
-		By("geting restart delay when capped")
-		var (
-			delay1 time.Duration
-			err    error
-		)
-		for i := 0; i < 3; i++ {
-			delay1, err = getRestartDelay(f.Client, pod, f.Namespace.Name, podName, containerName)
-			if err != nil {
-				framework.Failf("timed out waiting for container restart in pod=%s/%s", podName, containerName)
-			}
-
-			if delay1 < kubelet.MaxContainerBackOff {
-				continue
-			}
-		}
-
-		if (delay1 < kubelet.MaxContainerBackOff) || (delay1 > maxBackOffTolerance) {
-			framework.Failf("expected %s back-off got=%s in delay1", kubelet.MaxContainerBackOff, delay1)
-		}
-
-		By("getting restart delay after a capped delay")
-		delay2, err := getRestartDelay(f.Client, pod, f.Namespace.Name, podName, containerName)
-		if err != nil {
-			framework.Failf("timed out waiting for container restart in pod=%s/%s", podName, containerName)
-		}
-
-		if delay2 < kubelet.MaxContainerBackOff || delay2 > maxBackOffTolerance { // syncloop cumulative drift
-			framework.Failf("expected %s back-off got=%s on delay2", kubelet.MaxContainerBackOff, delay2)
 		}
 	})
